fix: point ws icons to correct folder
