+++ packages/client/src/client/App/LiveRates/Tile/TearOut/handleTearOut.openfin.ts
@@ -1,7 +1,7 @@
 import { ROUTES_CONFIG } from "@/client/constants"
 import { RT_FX_MAIN_WINDOW_NAME } from "@/client/OpenFin/utils/window"
 import { calculateWindowCoordinates, Offset } from "@/client/utils"
-import { constructUrl } from "@/client/utils/url"
+import { constructUrl } from "@/client/utils/constructUrl"
 import { openWindow } from "@/client/utils/window/openWindow"
 
 import { tearOut } from "./state"
