+++ packages/client/src/client/OpenFin/apps/Launcher/applicationConfigurations.ts
@@ -1,7 +1,7 @@
 import OpenFin from "@openfin/core"
+import { ENVIRONMENT } from "client/constants"
 
-import { ENVIRONMENT } from "@/client/constants"
-import { constructUrl } from "@/client/utils/url"
+import { constructUrl } from "@/client/utils/constructUrl"
 
 import {
   limitCheckerIcon,
