+++ packages/client/src/client/OpenFin/apps/LimitChecker/LimitCheckResults.tsx
@@ -9,7 +9,6 @@ import {
 import { LimitCheckStatus, LimitCheckTrade } from "@/services/trades/types"
 
 import { limitResult$ } from "./state"
-
 let tradeId = 0
 
 const tableRows$ = limitResult$.pipe(
