+++ packages/client/src/client/OpenFin/utils/templates/creditTemplateLayouts.ts
@@ -1,5 +1,6 @@
-import { ROUTES_CONFIG } from "@/client/constants"
-import { constructUrl } from "@/client/utils/url"
+import { ROUTES_CONFIG } from "client/constants"
+
+import { constructUrl } from "@/client/utils/constructUrl"
 
 import {
   createLayoutContent,
