+++ packages/client/src/workspace/home/utils.ts
@@ -216,8 +216,6 @@ export const handleAppSelection = async (
           dealtCurrency,
         } = appEntry as any
 
-        console.log("Action on execute", appEntry)
-
         await execute({
           currencyPair,
           spotRate,
